// src/modules/preacher/enums/preacher-search-type.enum.ts
export enum PreacherSearchType {
  FirstNames = 'first_names',
  LastNames = 'last_names',
  FullNames = 'full_names',
  BirthDate = 'birth_date',
  BirthMonth = 'birth_month',
  Gender = 'gender',
  MaritalStatus = 'marital_status',
  OriginCountry = 'origin_country',
  ZoneName = 'zone_name',
  AvailablePreachersByZone = 'available_preachers_by_zone',
  FamilyGroupCode = 'family_group_code',
  FamilyGroupName = 'family_group_name',
  ResidenceCountry = 'residence_country',
  ResidenceDepartment = 'residence_department',
  ResidenceProvince = 'residence_province',
  ResidenceDistrict = 'residence_district',
  ResidenceUrbanSector = 'residence_urban_sector',
  ResidenceAddress = 'residence_address',
  RecordStatus = 'record_status',
}

export const PreacherSearchTypeNames: Partial<Record<PreacherSearchType, string>> = {
  [PreacherSearchType.FirstNames]: 'Nombres',
  [PreacherSearchType.LastNames]: 'Apellidos',
  [PreacherSearchType.FullNames]: 'Nombres y Apellidos',
  [PreacherSearchType.ZoneName]: 'Nombre de Zona',
  [PreacherSearchType.FamilyGroupCode]: 'Código de grupo familiar',
  [PreacherSearchType.FamilyGroupName]: 'Nombre de grupo familiar',
  [PreacherSearchType.BirthDate]: 'Fecha de nacimiento',
  [PreacherSearchType.BirthMonth]: 'Mes de nacimiento',
  [PreacherSearchType.Gender]: 'Género',
  [PreacherSearchType.MaritalStatus]: 'Estado civil',
  [PreacherSearchType.OriginCountry]: 'País (origen)',
  [PreacherSearchType.ResidenceCountry]: 'País (residencia)',
  [PreacherSearchType.ResidenceDepartment]: 'Departamento (residencia)',
  [PreacherSearchType.ResidenceProvince]: 'Provincia (residencia)',
  [PreacherSearchType.ResidenceDistrict]: 'Distrito (residencia)',
  [PreacherSearchType.ResidenceUrbanSector]: 'Sector Urbano (residencia)',
  [PreacherSearchType.ResidenceAddress]: 'Dirección (residencia)',
  [PreacherSearchType.RecordStatus]: 'Estado de registro',
};